Added Order-Independent Sorting to Transparency
- Transparency now works as intended with up to 4 layers (This can be changed if needed)
- Objects with transparency now have to use different shader and renderPipelineState. This might be "fixed" in the future.

+++ Metal Engine#13/Graphics/RenderShaders/Shadows.metal
@@ -32,8 +32,8 @@ public:
         float bias = 0.0001;
         
         int index = 0;
-        for (int x = -halfSquareSample; x <= halfSquareSample; x++) {
-            for (int y = -halfSquareSample; y <= halfSquareSample; y++) {
+        for (short x = -halfSquareSample; x <= halfSquareSample; x++) {
+            for (short y = -halfSquareSample; y <= halfSquareSample; y++) {
                 float2 samplePosition =  float2(x, y) * texelSize;
                 samplePosition += samplePositionDefault;
                 float closestDepth = clamp(shadowMap1.sample(sampler2d, samplePosition), 0.0, 1.0);
